integrate month buttons to single destination query
